requirements.txt
ipykernel==6.29.5
matplotlib==3.10.0
openai==1.63.2
pandas==2.2.3
python-dotenv==1.0.1
pillow==11.1.0
SpeechRecognition
sounddevice
playsound==1.2.2
pyaudio